feat: 排盘与 JS iztro v2.5.8 完全对齐，金标测试转零容忍
排盘修正：
- 左辅右弼改用闰月修正后的月索引落宫（连带修好三台、八座、恩光、
  天贵的基准索引）
- 补上遗漏的华盖与天德两颗杂耀
- 旬空的阴阳判定改为「年支奇偶 vs 落宫索引」，而非原先的干支阴阳对比
- 晚子时（ti=12）紫微从次日起，跨月末回绕，与 dayDivide='forward' 一致
- 四柱 chinese_date 重写：日柱取精确值（晚子时进位次日），月柱走五虎遁
  且年份按正月初一分界，各语言格式与 iztro 一致
- 时辰、星座、生肖补齐六语言词表（13 时辰 / 12 星座 / 12 生肖）

测试：
- tier1/tier2/matrix 移除全部容差列表与 ti=12 跳过项，
  780 + 37,440 + matrix 例一律断言完全相等
- 重新生成 tier3 自检基线（284,895 例）
- 金标生成器锁定 iztro 2.5.8 精确版本
- 裸指针 FFI 导出标注 unsafe，clippy 零告警

+++ src/astro/palace.rs
@@ -131,8 +131,8 @@ pub fn get_decadals_and_ages(
     for i in 0..12 {
         let ages_for_palace: Vec<u32> = (0..10).map(|j| (12 * j + i + 1) as u32).collect();
         let idx = match gender {
-            Gender::Male => fix_index(age_idx as i32 + i as i32, 12),
-            Gender::Female => fix_index(age_idx as i32 - i as i32, 12),
+            Gender::Male => fix_index(age_idx as i32 + i, 12),
+            Gender::Female => fix_index(age_idx as i32 - i, 12),
         };
         final_ages[idx] = ages_for_palace;
     }
